more upload cred related code
