+++ core/src/main/scala/org/apache/spark/scheduler/cluster/CoarseGrainedClusterMessage.scala
@@ -58,6 +58,8 @@ private[spark] object CoarseGrainedClusterMessages {
 
   case class StopSparkContext() extends CoarseGrainedClusterMessage
 
+  case class DelegateCredentials(credentials: Array[Byte]) extends CoarseGrainedClusterMessage
+
   // Executors to driver
   case class RegisterExecutor(
       executorId: String,
